fix: complete persona migration by removing deprecated VRM routes and SDK code
- Update `input.parseMenu()` to expect `linkedPersona` (string) and return `Persona`
- Remove deprecated `linkedVrm` fields, `SetLinkedVrmRequest`, and VRM link methods from SDK webviews
- Remove entire `/vrm/*` route tree from HTTP server (persona routes cover all functionality)
- Move `PlayBody`/`Repeat` types into `persona/vrm_ops.rs` to decouple from removed vrm module
- Regenerate OpenAPI spec
