--- engine/crates/homunculus_http_server/src/route/vrm.rs
@@ -1,18 +1,0 @@
-//! `/vrm` provides methods for interacting with VRM models.
-
-pub mod bone;
-pub(crate) mod despawn;
-pub(crate) mod events;
-pub mod expressions;
-pub(crate) mod get;
-pub mod look;
-pub mod persona;
-pub mod position;
-pub mod snapshot;
-pub(crate) mod spawn;
-pub mod speech;
-pub mod spring_bones;
-pub mod state;
-pub(crate) mod stream;
-pub mod vrma;
-pub(crate) mod wait_load;
