--- engine/crates/homunculus_http_server/src/route/vrm/look.rs
@@ -1,116 +1,0 @@
-use crate::extract::EntityId;
-use axum::extract::{Path, State};
-use bevy::prelude::Entity;
-use homunculus_api::prelude::axum::{HttpResult, IntoHttpResult};
-use homunculus_api::vrm::VrmApi;
-
-/// Disable look-at control for the specified VRM entity.
-#[utoipa::path(
-    delete,
-    path = "/look",
-    tag = "vrm",
-    params(("entity" = String, Path, description = "Entity ID")),
-    responses(
-        (status = 200, description = "Look-at disabled"),
-        (status = 404, description = "Entity not found"),
-    ),
-)]
-pub async fn unlook(State(api): State<VrmApi>, EntityId(vrm): EntityId) -> HttpResult {
-    api.unlook(vrm).await.into_http_result()
-}
-
-/// Set look-at target to another entity.
-#[utoipa::path(
-    put,
-    path = "/look/target/{target}",
-    tag = "vrm",
-    params(
-        ("entity" = String, Path, description = "VRM entity ID"),
-        ("target" = String, Path, description = "Target entity ID"),
-    ),
-    responses(
-        (status = 200, description = "Look-at target set"),
-        (status = 404, description = "Entity not found"),
-    ),
-)]
-pub async fn target(
-    State(api): State<VrmApi>,
-    Path((vrm, target)): Path<(Entity, Entity)>,
-) -> HttpResult {
-    api.look_at_target(vrm, target).await.into_http_result()
-}
-
-/// Set look-at to follow the cursor.
-#[utoipa::path(
-    put,
-    path = "/look/cursor",
-    tag = "vrm",
-    params(("entity" = String, Path, description = "Entity ID")),
-    responses(
-        (status = 200, description = "Look-at cursor mode set"),
-        (status = 404, description = "Entity not found"),
-    ),
-)]
-pub async fn cursor(State(api): State<VrmApi>, EntityId(vrm): EntityId) -> HttpResult {
-    api.look_at_cursor(vrm).await.into_http_result()
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::tests::{call, test_app};
-    use bevy::prelude::*;
-    use bevy_vrm1::prelude::LookAt;
-
-    #[tokio::test]
-    async fn test_unlook() {
-        let (mut app, router) = test_app();
-        let entity = app.world_mut().spawn(LookAt::Cursor).id();
-        app.update();
-
-        let request = axum::http::Request::delete(format!("/vrm/{}/look", entity.to_bits()))
-            .body(axum::body::Body::empty())
-            .unwrap();
-        call(&mut app, router, request).await;
-        assert!(
-            !app.world().entity(entity).contains::<LookAt>(),
-            "LookAt component should be removed"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_look_at_target() {
-        let (mut app, router) = test_app();
-        let entity = app.world_mut().spawn(LookAt::Cursor).id();
-        let target = app.world_mut().spawn_empty().id();
-        app.update();
-
-        let request = axum::http::Request::put(format!(
-            "/vrm/{}/look/target/{}",
-            entity.to_bits(),
-            target.to_bits()
-        ))
-        .body(axum::body::Body::empty())
-        .unwrap();
-        call(&mut app, router, request).await;
-        assert_eq!(
-            app.world().entity(entity).get::<LookAt>().unwrap(),
-            &LookAt::Target(target),
-        );
-    }
-
-    #[tokio::test]
-    async fn test_look_at_cursor() {
-        let (mut app, router) = test_app();
-        let entity = app.world_mut().spawn(LookAt::Cursor).id();
-        app.update();
-
-        let request = axum::http::Request::put(format!("/vrm/{}/look/cursor", entity.to_bits(),))
-            .body(axum::body::Body::empty())
-            .unwrap();
-        call(&mut app, router, request).await;
-        assert_eq!(
-            app.world().entity(entity).get::<LookAt>().unwrap(),
-            &LookAt::Cursor,
-        );
-    }
-}
